+++ .env.example
@@ -1,0 +1,4 @@
+TENDERLY_ACCESS_KEY
+TENDERLY_PROJECT_SLUG
+DEVNET_RPC_URL
+ETHERSCAN_API_KEY
