+++ remappints.txt
@@ -1,0 +1,14 @@
+ERC4626/=lib/properties/lib/ERC4626/contracts/
+ds-test/=lib/forge-std/lib/ds-test/src/
+erc4626-tests/=lib/openzeppelin-contracts-upgradeable/lib/erc4626-tests/
+erc721m/=lib/erc721m/contracts/
+forge-gas-snapshot/=lib/forge-gas-snapshot/src/
+forge-std/=lib/forge-std/src/
+hardhat-deploy/=node_modules/hardhat-deploy/
+hardhat/=node_modules/hardhat/
+openzeppelin-contracts-upgradeable/=lib/openzeppelin-contracts-upgradeable/
+@openzeppelin/contracts/=lib/openzeppelin-contracts/contracts/
+openzeppelin/=lib/openzeppelin-contracts-upgradeable/contracts/
+properties/=lib/properties/contracts/
+solady/=lib/solady/
+solmate/=lib/solmate/src/
