+++ src/tokens/ERC721A.sol
@@ -2,7 +2,7 @@
 // ERC721A Contracts v4.2.3
 // Creator: Chiru Labs
 
-pragma solidity ^0.8.4;
+pragma solidity ^0.8.20;
 
 import "../interfaces/IERC721A.sol";
